+++ application/config/bancha.php
@@ -61,7 +61,7 @@ $config['admin_languages'] = array(
 /*
  * Framework version
  */
-define('BANCHA_VERSION', '0.7.8');
+define('BANCHA_VERSION', '0.7.9');
 
 /*
  * Framework name