+++ application/models/bancha/model_settings.php
@@ -52,8 +52,6 @@ Class Model_settings extends CI_Model
 	 */
 	public function set($key, $val, $module = 'general')
 	{
-		$module = strtolower($module);
-
 		//Does this setting already exists?
 		$exists = isset($this->_items[$module][$key]);
 
@@ -81,6 +79,31 @@ Class Model_settings extends CI_Model
 						array('name' => $key, 'value' => $val, 'module' => $module)
 				   );
 		}
+		return FALSE;
+	}
+
+	/**
+	 * Create/updates a blocks of a theme template
+	 * @param string $key
+	 * @param mixed $val
+	 * @param string $theme
+	 * @param string $template
+	 */
+	public function set_block($key, $val, $theme, $template = 'default')
+	{
+		return $this->set($key, $val, 'blocks-' . $theme . '-' . $template);
+	}
+
+	/**
+	 * Returns a single block of a theme template
+	 * @param string $key
+	 * @param string $theme
+	 * @param string $template
+	 * @return mixed value
+	 */
+	public function get_block($key, $theme, $template = 'default')
+	{
+		return $this->get($key, 'blocks-' . $theme . '-' . $template);
 	}
 
 	/**
@@ -91,7 +114,6 @@ Class Model_settings extends CI_Model
 	 */
 	public function get($key, $module = 'general')
 	{
-		$module = strtolower($module);
 		if (isset($this->_items[$module][$key]))
 		{
 			return $this->_items[$module][$key];